* V5-7-patches:
  Rework commit b4a19505a05816e94e01de97bf99ac37fcdc6d23.

+++ agent/mibgroup/disman/schedule/schedCore.c
@@ -301,15 +301,14 @@ sched_nextTime( struct schedTable_entry *entry )
         } else {
              entry->schedNextRun = now + entry->schedInterval;
         }
-        DEBUGMSGTL(("disman:schedule:time", "periodic: (%lld) %s",
-                    (long long) entry->schedNextRun,
-                    ctime(&entry->schedNextRun)));
+        DEBUGMSGTL(("disman:schedule:time", "periodic: (%ld) %s",
+                    (long) entry->schedNextRun, ctime(&entry->schedNextRun)));
         break;
 
     case SCHED_TYPE_ONESHOT:
         if ( entry->schedLastRun ) {
-            DEBUGMSGTL(("disman:schedule:time", "one-shot: expired (%lld) %s",
-                        (long long) entry->schedNextRun,
+            DEBUGMSGTL(("disman:schedule:time", "one-shot: expired (%ld) %s",
+                        (long) entry->schedNextRun,
                         ctime(&entry->schedNextRun)));
             return;
         }
@@ -424,9 +423,8 @@ sched_nextTime( struct schedTable_entry *entry )
          * 'next_tm' now contains the time for the next scheduled run
          */
         entry->schedNextRun = mktime( &next_tm );
-        DEBUGMSGTL(("disman:schedule:time", "calendar: (%lld) %s",
-                    (long long) entry->schedNextRun,
-                    ctime(&entry->schedNextRun)));
+        DEBUGMSGTL(("disman:schedule:time", "calendar: (%ld) %s",
+                    (long) entry->schedNextRun, ctime(&entry->schedNextRun)));
         return;
 
     default:
